Moved reactive query gateway to separate module.
