+++ reactive/src/test/java/org/axonframework/reactive/DefaultReactiveQueryGatewayTest.java
@@ -14,10 +14,15 @@
  * limitations under the License.
  */
 
-package org.axonframework.queryhandling;
+package org.axonframework.reactive;
 
 import org.axonframework.common.Registration;
 import org.axonframework.messaging.MessageDispatchInterceptor;
+import org.axonframework.queryhandling.GenericQueryResponseMessage;
+import org.axonframework.queryhandling.QueryBus;
+import org.axonframework.queryhandling.QueryMessage;
+import org.axonframework.queryhandling.QueryResponseMessage;
+import org.axonframework.queryhandling.UpdateHandler;
 import org.axonframework.queryhandling.responsetypes.ResponseTypes;
 import org.junit.*;
 import org.mockito.*;
