+++ spring-boot-autoconfigure/src/main/java/org/axonframework/boot/autoconfig/AxonAutoConfiguration.java
@@ -39,11 +39,11 @@ import org.axonframework.messaging.SubscribableMessageSource;
 import org.axonframework.messaging.correlation.CorrelationDataProvider;
 import org.axonframework.messaging.correlation.MessageOriginProvider;
 import org.axonframework.messaging.interceptors.CorrelationDataInterceptor;
-import org.axonframework.queryhandling.DefaultReactiveQueryGateway;
 import org.axonframework.queryhandling.QueryBus;
 import org.axonframework.queryhandling.QueryInvocationErrorHandler;
-import org.axonframework.queryhandling.ReactiveQueryGateway;
 import org.axonframework.queryhandling.SimpleQueryBus;
+import org.axonframework.reactive.DefaultReactiveQueryGateway;
+import org.axonframework.reactive.ReactiveQueryGateway;
 import org.axonframework.serialization.*;
 import org.axonframework.serialization.json.JacksonSerializer;
 import org.axonframework.serialization.xml.XStreamSerializer;
